requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
httpx==0.25.0